Tidied up module handling
