+++ internal/app/tfsec/parser/load_module_metadata.go
@@ -1,0 +1,38 @@
+package parser
+
+import (
+	"encoding/json"
+	"os"
+	"path/filepath"
+)
+
+type ModulesMetadata struct {
+	Modules []ModuleMetadata `json:"Modules"`
+}
+
+type ModuleMetadata struct {
+	Key     string `json:"Key"`
+	Source  string `json:"Source"`
+	Version string `json:"Version"`
+	Dir     string `json:"Dir"`
+}
+
+func LoadModuleMetadata(fullPath string) (*ModulesMetadata, error) {
+	metadataPath := filepath.Join(fullPath, ".terraform/modules/modules.json")
+	if _, err := os.Stat(metadataPath); err != nil {
+		return nil, err
+	}
+
+	f, err := os.Open(metadataPath)
+	if err != nil {
+		return nil, err
+	}
+	defer func() { _ = f.Close() }()
+
+	var metadata ModulesMetadata
+	if err := json.NewDecoder(f).Decode(&metadata); err != nil {
+		return nil, err
+	}
+
+	return &metadata, nil
+}
